requirements.txt
# Core Django dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# LLM Integration
openai==1.12.0
transformers==4.37.2
torch==2.2.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
markdown==3.5.2

# Vector Database & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.1
numpy==1.26.3
tiktoken==0.5.2

# Caching
django-redis==5.4.0
redis==5.0.1

# Development
python-decouple==3.8
